pyproject.toml
[project]
name = "swarm-research"
version = "0.1.0"
description = "Autonomous multi-agent optimization framework"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "openai>=1.50",
    "python-dotenv>=1.0",
]
